requirements.txt
discord.py==2.4.0
requests==2.32.3
flask==3.0.3
gunicorn==23.0.0
psycopg2-binary==2.9.9